+++ recipes/cleaner.rb
@@ -24,7 +24,7 @@ require 'fileutils'
   # Next is needed to prevent resource definition when it is not needed.
   # `only_if` statement blocks notifications which we are sending
   next unless node['init_package'] == 'systemd' and
-              %w(static_pods upstart).include? node['kubernetes']['install_via']
+    %w(static_pods upstart).include? node['kubernetes']['install_via']
 
   systemd_unit "kube-#{srv}.service" do
     action [:disable, :stop, :delete]
@@ -131,4 +131,20 @@ end
 file '/etc/systemd/network/kubernetes_services.network' do
   action :delete
   notifies :restart, 'service[systemd-networkd]'
+end
+
+# Cleanup dashboard manifests
+%w(
+  dashboard-sa
+  dashboard-role
+  dashboard-rolebinding
+  dashboard-deployment
+  dashboard-svc
+).each do |srv|
+  file "/etc/kubernetes/addons/#{srv}.yaml" do
+    action :delete
+    only_if do
+      %(systemd upstart).include? node['kubernetes']['install_via']
+    end
+  end
 end